Optimize admin dashboard and user navigation
